Remove author tags and format Javadocs. Rename `DbTableNameFactory` to `TableNames.`
