+++ rdbms/src/main/java/io/spine/server/storage/jdbc/aggregate/InsertEventCountQuery.java
@@ -46,7 +46,6 @@ class InsertEventCountQuery<I> extends WriteEventCountQuery<I> {
         return new Builder<>();
     }
 
-    @SuppressWarnings("ClassNameSameAsAncestorName")
     static class Builder<I> extends WriteEventCountQuery.Builder<Builder<I>,
                                                                  InsertEventCountQuery<I>,
                                                                  I> {
