+++ rdbms/src/main/java/io/spine/server/storage/jdbc/query/AbstractSelectByIdQuery.java
@@ -25,13 +25,12 @@ import com.querydsl.core.types.dsl.BooleanExpression;
 
 /**
  * An abstract base for the queries which read a single record by an ID.
- *
- * @author Dmytro Grankin
  */
-public abstract class AbstractSelectByIdQuery<I, R> extends IdAwareQuery<I> implements SelectQuery<R> {
+public abstract class AbstractSelectByIdQuery<I, R>
+        extends IdAwareQuery<I> implements SelectQuery<R> {
 
-    protected AbstractSelectByIdQuery(Builder<I, ? extends Builder, ? extends StorageQuery> builder) {
-        super(builder);
+    protected AbstractSelectByIdQuery(Builder<I, ? extends Builder, ? extends StorageQuery> b) {
+        super(b);
     }
 
     /**
